[FEAT&REFACT]: Refacotring Recruit's components
